Fix ascii encoding bug in camera.annotate_text
This fixes the bug that was preventing the sensor data from being
annotated onto the photos.

The built in camera.annotate_text method uses the ascii codec to encode
text, and write it onto the images. However, before this fix, the
'micro' symbol, which isn't a part of the ascii codec, was in the text.
This caused the error, and has now been corrected to 'micro' (the word).