Integrate Atlas sensors code (#18)
* Add atlas sensor library files and sensor driver code

* Integrate Atlas sensor classes.

* Fix faulty get data methods.

* Fix faulty _get_data method.

* Add docstring, comments.

* Minor changes.

* Fix remote I/O error with factory reset method.

* Fix initialise sensor method.

* Add ABC for AtlasI2C base class.

* Resolve TODO's.

* Minor change.

* Rename AtlasI2C.py to atlasI2C.py

* Add minor type hinting improvements.

* Change incorrect attribute names, add method calls

* Fix import

* Fix incorrect method name

* Add initialise sensor method

* Minor edit

* Add sensor annotations on image

* Fix bugs in sensor code

* Add correction to the name of the keys in sensor readings

* Fix sensor reading for percentage oxygen

* Try to fix string to float conversion error

* Remove duplicate calibration attempts

* Try to fix list index error

* Add testfile for bug fixes

* Add additional checks for bugs

* Minor change to follow convention

* Add print statement to catch bug

* Change method from public to private

* Add TODO note for later improvement

* Fix import error and incorrect method name

* Add additional checks for bugs

* Add further checks for bugs

* Make minor changes

* Add further checks for list index bug

* Fix parameterEnableError for EC_Sensor

Along with correcting incorrect syntax, add more wait time to ensure
that previous calls are processed completely.

* AdD TODO notes for later dev + correct key names

To find the cause of the error of the data not being annotated to the
photos, the names of the keys have all been made uniform.

Seperately, TODO comments have been added where additional features need
to be added for the purpose of uniformity.

* Add more TODO comments

* Add checks to find missing photo annotation bug

The sensors were being read, but the readings aren't being annotated on
the photos at all. The checks will now show whether there is a prob with
data going in/out of the annotate_text_string method.

* Delete testfile.py

This file was required only to test the sensors, to find and solve the bug preventing them from being read properly.
This bug has since been found and fixed.

* Remove testfile.py

* Fix ascii encoding bug in camera.annotate_text

This fixes the bug that was preventing the sensor data from being
annotated onto the photos.

The built in camera.annotate_text method uses the ascii codec to encode
text, and write it onto the images. However, before this fix, the
'micro' symbol, which isn't a part of the ascii codec, was in the text.
This caused the error, and has now been corrected to 'micro' (the word).

* Remove testing print statements

* Remove testing print statements

* Remove testing print statements

* Remove _UNITS class attribute from all classes

The _UNITS class attribute was required during testing, but is no longer
required. The units of each measurement parameter can now be found in
the docstrings of each class.

* Remove get_header_row method from classes

This method was required during testing, no longer required.

* Refactor get_data methods for all classes

Print statements in the methods that were required for testing have been
removed.
Also, the methods now return -1 in the event of an exception, to be
consistent with the rest of the code.

* Remove print statements from DO_Sensor init

* Make minor changes

* Make minor change

* Add docstring for annotate_text_string method

* Remove unaccessed variable from capture_video

* Remove duplicate method definition

* Make a minor formatting change

* Fix syntax error

* Add type annotations

* Add exception handling classes for Atlas sensors

* Make a minor a change

* Add TODO for code review

* Add type annotations

* Fix missing import

* Resolve TODO comment

* Correct docstring in EC_Sensor

* Add missing units in annotate_text_string

The units for some of the readings from the sensors were missing. These
have now been added to provide clarity when looking at the readings.

* Remove extra line space

* Add TODO comment

* Remove unaccessed imports

* Add type annotations for init methods

* Correct type annotation for sensor_data

* Remove redundant imports

* Correct docstring for annotate_text_string

* Minorly reformat code

* Resolve TODO comment, remove print statement

* Remove unaccessed code

* Improve error/exception  message in _set_cal_data

* Refactor code a little

* Remove confusing comments

* Add Atlas sensor imports sensor package init

* Add type annotations

* Add TODO comment

* Fix write_sensor_data sensor object with timestamp

Co-authored-by: Milind220 <68847270+Milind220@users.noreply.github.com>